marimo-dev/pyproject.toml
[project]
name = "marimo-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.11.0",
    "influxdb>=5.3.2",
    "influxdb-client>=1.50.0",
    "marimo>=0.20.4",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "redis>=7.3.0",
]
